profile: add rudimentary cross-compilation support
This commit adds a hacky attempt at cross-compilation support
through build "profiles", which are basically just collections
of architecture-specific compiler flags or settings.

This is a very incomplete implementation and likely not the best
solution, largely wrote this to simplify the large amount of OPTION='s
that were getting unwieldy to remember all of.