imgui: always allow controller navigation if option is set
Currently, there are two issues with pure-controller navigation:
 - there is no gamepad button assigned to open/close the menu
 - controller navigation is only enabled when *opening* the menu

This is a bit of a chicken-and-egg problem. Adding a button to open
the menu still requires controller navigation to be enabled to actually
do anything. This means when the application launches, the user has
to press F1 on a keyboard to open/reopen the menu before the gamepad
can be used in menus.

This simply enables keyboard/gamepad support at init-time, and removes
the code that enables/disables gamepad support on menu open/close.